SAK-11544
Commit of some test code - please don't look at this - it is a bit of refactor in Mail Action
and a bad hack to fill up a mail archive by pressing the options button.  I just wanted to put
this in my happy little branch so I did not have to keep carrying this code around manually.



git-svn-id: https://source.sakaiproject.org/svn/mailarchive/branches/SAK-11544@39631 66ffb92e-73f9-0310-93c1-f5514f145a0a