SAK-11544
Performance improve scenarios where you are searching and paging.  Allow the 
SqlReader to return "this" to indicate that the Reader has readed the 
end of the record set.  The SqlService DBRead stops reading at that point
saving the need to transfer lots of data from the DB server just to 
throw it away.  

In paricular because usually folks only page through the first few 
pages anyways this wil make the common case pretty fast when there
are a lot of hits.  When there are a few hits - the tunable higher 
level message list caching will help there.

What is still slow (but tolerable) is the page to the last page of 
a list when searching.  Ultimately this means a full scan of the 
result set to find those last 20 messages.

This adds a requirement to have a new version SqlService.

For now this new code is here:

https://source.sakaiproject.org/svn/db/branches/SAK-11544/

Code is added to detect and cope with the older 
version of SqlService.  This code to process the list 
right after dbRead returns is and look for instances 
of SqlReaders is there in BaseDbDoubleStorage and it
can be removed once this moves toward trunk and we are sure
of the version of SqlService.



git-svn-id: https://source.sakaiproject.org/svn/mailarchive/branches/SAK-11544@40876 66ffb92e-73f9-0310-93c1-f5514f145a0a