SAK-13584
Add feature to suppress the search option if there are too many messages
in the channel.

sakai.mailbox.search-threshold=2500

The default is 2500 messages - above this number searches get somewhat slow as they 
use like clauses in the newly exploded columns.  This is *different* from the Sakai 2-5
version of Mail Tool - it actually tried to search large corpuses but features were
suppressed when the search result returned too many results.  The Sakai 2.5 approach 
actually retrieved and deserialized all of the data in the App Server.

Some testing can be done to determine if this is a good limit.  Some initial testing
of a 25000 message corpus suggested that it takes about 1 second per 1000 messages doing 
a like clause.  That testing was done in MySql with a prety weak system and slow disk
on a cunsumer desktop - so perhaps this threshold can be higher when the server is
more performant.  Some basic like queries can be done on the MAILARCHIVE_MESSAGE for 
a channel with a large corpus to hep determine the right setting for this parameter 
for a particular database server.



git-svn-id: https://source.sakaiproject.org/svn/mailarchive/branches/SAK-13584@49413 66ffb92e-73f9-0310-93c1-f5514f145a0a